datagen/requirements.txt
Faker>=33.0.0
pandas
confluent-kafka
psycopg2-binary>=2.9.10